+++ qq/errors.py
@@ -8,3 +8,15 @@ class ApplicationNotStartedError(QQError):
 
 class AlreadyStartedError(QQError):
     pass
+
+
+class PluginLacksOfKeyError(QQError):
+    pass
+
+
+class WrongKeyForPluginError(QQError):
+    pass
+
+
+class PluginKeyAlreadyDefinedError(QQError):
+    pass