+++ qq/plugins/tests/test_settings.py
@@ -11,6 +11,7 @@ from qq import Application
 from qq import Context
 from qq.plugins.settings import PrefixedStringsDict
 from qq.plugins.settings import SettingsBasedPlugin
+from qq.plugins.settings import SettingsInjector
 from qq.plugins.settings import SettingsPlugin
 from qq.plugins.settings import _import
 from qq.testing import PluginFixtures
@@ -138,3 +139,20 @@ class TestSettingsBasedPlugin:
 
     def test_get_my_settings_when_context(self, context, plugin):
         assert plugin.get_my_settings(context) == sentinel.settings
+
+
+class TestSettingsInjector:
+    PLUGINKEY = "ksdyuwyta"
+
+    @fixture
+    def app(self):
+        app = Application()
+        app.globals["settings"] = {self.PLUGINKEY: sentinel.settings}
+        return app
+
+    def test_flow(self, app):
+        context = {
+            SettingsPlugin.key: {self.PLUGINKEY: sentinel.pluginsettings}
+        }
+        result = SettingsInjector.fun(context=context, key=self.PLUGINKEY)
+        assert result == sentinel.pluginsettings