+++ docsrc/docs/CHANGELOG.rst
@@ -7,6 +7,16 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <http://keepachangelog.com/>`_ and this project adheres to [Semantic Versioning]
 (http://semver.org/).
 
+Version: 1.0.5
+==============
+
+* Added
+    Add Plugin validation.
+    Add Singleton Plugins functionality.
+    Add Plugin Container in Application.
+* Changed
+    Use Plugin Container instead of OrderedDict for plugins.
+
 Version: 1.0.4
 ==============